refactor: extract shorten router from server.ts
